update to match new interface
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@14321 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ nmad/protocols/proto_mpi/include/mpi_types.h
@@ -54,7 +54,7 @@ typedef int MPI_Request_type;
 
 typedef struct MPI_Request_s {
   MPI_Request_type request_type;
-  unsigned long request_id;
+  intptr_t request_id;
 } MPI_Request_t;
 
 typedef struct MPI_Request_s *MPI_Request;
